requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
docopt==0.6.2
fastapi==0.116.1
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-ollama==0.3.7
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-checkpoint-redis==0.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.20
ml_dtypes==0.5.3
numpy==2.3.2
ollama==0.5.3
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
ply==3.11
proto-plus==1.26.1
protobuf==6.32.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ulid==3.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
redisvl==0.8.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
xxhash==3.5.0
yarg==0.1.10
zstandard==0.24.0
